Define color palette CSS variables
Instead of just manually setting the colors every single time I will
need them, I would rather just call variables instead, so that if I want
to change the color scheme later on, that it is much simpler to do so.

This will require me to use the :root pseudo-class and custom properties
both of which are concepts not introduced in the TOP curriculum at this
point but I happen to recall existing.